Change compiler rules hardcoding
The compiler rules are no longer hardcoded but instead stored in a vector. Each rule has a type, a match function and an apply function. This allows rules to be only applied when both the type (Optional) and match function are correct

Adding a new rule is as simple as inheriting from the compiler_rule class, implementing the methods and adding it to the rules vector.